+++ src/main/java/dinkplugin/notifiers/data/PetNotificationData.java
@@ -1,0 +1,24 @@
+package dinkplugin.notifiers.data;
+
+import dinkplugin.message.Field;
+import lombok.EqualsAndHashCode;
+import lombok.Value;
+
+import java.util.Collections;
+import java.util.List;
+
+@Value
+@EqualsAndHashCode(callSuper = false)
+public class PetNotificationData extends NotificationData {
+    String petName;
+
+    @Override
+    public List<Field> getFields() {
+        if (petName == null || petName.isEmpty())
+            return super.getFields();
+
+        return Collections.singletonList(
+            new Field("Name", "```\n" + petName + "\n```")
+        );
+    }
+}